add reset password test, some test refactoring
